Move map generator scripts in Scripts/Map
