--- Assets/Editor/MapGeneratorEditor.cs
@@ -1,22 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEditor;
-
-[CustomEditor (typeof(MapGenerator))]
-public class MapGeneratorEditor : Editor
-{
-    public override void OnInspectorGUI()
-    {
-        MapGenerator mapGen = (MapGenerator)target;
-
-        if(DrawDefaultInspector() && mapGen.autoUpdate)
-        {
-            mapGen.DrawMapInEditor();
-        }
-        if(GUILayout.Button("Generate"))
-        {
-            mapGen.DrawMapInEditor();
-        }
-    }
-}
