+++ Assets/Scripts/Map/MapDisplay.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MapDisplay : MonoBehaviour
+{
+    public Renderer textureRenderer;
+
+    public void DrawTexture(Texture2D texture)
+    {
+            Debug.Log(textureRenderer);
+            textureRenderer.sharedMaterial.mainTexture = texture;
+            textureRenderer.transform.localScale = new Vector3(texture.width, 1, texture.height);
+        
+    }
+}
