--- Assets/scripts/MapDisplay.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MapDisplay : MonoBehaviour
-{
-    public Renderer textureRenderer;
-
-    public void DrawTexture(Texture2D texture)
-    {
-            Debug.Log(textureRenderer);
-            textureRenderer.sharedMaterial.mainTexture = texture;
-            textureRenderer.transform.localScale = new Vector3(texture.width, 1, texture.height);
-        
-    }
-}
