--- Assets/scripts/Noise.cs
@@ -1,88 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public static class Noise
-{
-    public enum NormalizeMode { Local, Global }
-    public static float[,] GenerateNoiseMap(int width, int height,int seed, float scale,int octaves, float persistance, float lacunarity,Vector2 offset, NormalizeMode normalizeMode)
-    {
-        float[,] noiseMap = new float[width, height];
-
-        System.Random prng = new System.Random(seed);
-        Vector2[] octaveOffsets = new Vector2[octaves];
-        float maxPossibleHeight = 0;
-        float amplitude = 0;
-        float frequency;
-
-        for (int i = 0; i < octaves; i++)
-        {
-            float offsetX = -(prng.Next(-100000,100000)+offset.x);
-            float offsetY = -(prng.Next(-100000,100000)+offset.y);
-           octaveOffsets[i] = new Vector2(offsetX, offsetY);
-
-            maxPossibleHeight += amplitude;
-            amplitude += persistance;
-        }
-
-        if(scale<=0)
-        {
-            scale = 0.00001f;
-        }
-
-        float maxLocalNoiseHeight = float.MaxValue;
-        float minLocalNoiseHeight = float.MinValue;
-
-        float halfWidth = width / 2f;
-        float halfHeight = height / 2f;
-        for (int x = 0; x < width; x++)
-        {
-            for (int y = 0; y < height; y++)
-            {
-                 amplitude = 1;
-                 frequency = 1;
-                float noiseHeight = 0;
-
-                for (int i = 0; i < octaves; i++)
-                {
-                    float sampleX = (x-halfWidth + octaveOffsets[i].x)  / scale*frequency ;
-                    float sampleY = (y-halfHeight + octaveOffsets[i].y)  / scale*frequency ;
-
-                    float perlinValue = Mathf.PerlinNoise(sampleX, sampleY)*2-1;
-                    noiseHeight += perlinValue*amplitude;
-                    amplitude *= persistance;
-                    frequency *= lacunarity;
-                }
-
-                if(noiseHeight<maxLocalNoiseHeight)
-                {
-                    maxLocalNoiseHeight = noiseHeight;
-                }
-                else if (noiseHeight> minLocalNoiseHeight)
-                {
-                    minLocalNoiseHeight = noiseHeight;
-                }
-
-                noiseMap[x, y] = noiseHeight;
-            }
-        }
-
-        for (int x = 0; x < width; x++)
-        {
-            for (int y = 0; y < height; y++)
-            {
-                if (normalizeMode == NormalizeMode.Local)
-                {
-                    noiseMap[x, y] = Mathf.InverseLerp(minLocalNoiseHeight, maxLocalNoiseHeight, noiseMap[x, y]);
-                }
-                else
-                {
-                    float normalizeHeight = (noiseMap[x, y] + 1) / (2f*maxPossibleHeight/3.5f);
-                    noiseMap[x, y] = normalizeHeight;
-                }
-            }
-        }
-
-        return noiseMap;
-    }
-}
